pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["selver96 <selver.said01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
djangorestframework-simplejwt = "4.8.0"
PyJWT = "^2.3.0"
pdbpp = "^0.10.3"
django-cors-headers = "^3.11.0"
PyMySQL = "^1.0.2"
Pillow = "^9.0.1"
django-ckeditor = "^6.2.0"
django-debug-toolbar = "^3.2.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
